[KYUUBI #5555][AUTHZ] Extractor common authorization rule class
